+++ test/events/eventCameraShake.test.js
@@ -5,10 +5,10 @@ test("Should be able to shake camera for a second", () => {
   compile(
     {
       shakeDirection: "diagonal",
-      time: 1
+      time: 1,
     },
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
   expect(mockCameraShake).toBeCalledWith(true, true, 60);
@@ -19,7 +19,7 @@ test("Should shake camera for half a second if time not set", () => {
   compile(
     {},
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
   expect(mockCameraShake).toBeCalledWith(true, false, 30);
@@ -29,10 +29,10 @@ test("Should shake camera horizontally if directions not set", () => {
   const mockCameraShake = jest.fn();
   compile(
     {
-      time: 1
+      time: 1,
     },
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
   expect(mockCameraShake).toBeCalledWith(true, false, 60);
@@ -43,10 +43,10 @@ test("Should shake camera horizontally if direction set to horizontal", () => {
   compile(
     {
       shakeDirection: "horizontal",
-      time: 1
+      time: 1,
     },
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
   expect(mockCameraShake).toBeCalledWith(true, false, 60);
@@ -57,10 +57,10 @@ test("Should shake camera vertically if direction set to vertical", () => {
   compile(
     {
       shakeDirection: "vertical",
-      time: 1
+      time: 1,
     },
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
   expect(mockCameraShake).toBeCalledWith(false, true, 60);
@@ -71,12 +71,11 @@ test("Should be able to shake camera for one and a half seconds", () => {
   compile(
     {
       shakeDirection: "diagonal",
-      time: 1.5
+      time: 1.5,
     },
     {
-      cameraShake: mockCameraShake
+      cameraShake: mockCameraShake,
     }
   );
-  expect(mockCameraShake).toBeCalledWith(true, true, 60);
-  expect(mockCameraShake).toBeCalledWith(true, true, 30);
+  expect(mockCameraShake).toBeCalledWith(true, true, 90);
 });